Remove SearchRepositoryWithCountJdbi and SearchRepositoryJdbi
SearchRepositoryWithCountJdbi was added to solve a problem with
AbstractSearchRepositoryWithCount: that it required both the
SearchRepository and SearchRepositoryWithCount interfaces, which meant
that users had to implement both a `search` and a `searchWithCount`
method for their search queries. This has led to issues such as in
Tomra's Event Log, where the `search` method is left as
`TODO("Not yet implemented")`. SearchRepositoryWithCountJdbi solved this
by adding an additional layer of abstraction.

The problem with this additional layer of abstraction is the distance it
created from the implementation of search queries to their actual
execution. To see how a search query was executed, one would now have to
go to the sqlWhere field on the search query, then trace its use through
to the SearchRepositoryWithCountJdbi, then through the getByPredicate
method. This flow is different from how we normally use search
repositories, and is also inflexible in that the implementation is done
through fields on the search query class instead of in code.

Because of these issues, this commit removes the
SearchRepositoryWithCountJdbi, and aims to solve the root of the
problem: making AbstractSearchRepositoryWithCount no longer require the
SearchRepository interface, so that one only needs to implement the
search method with count. We also remove SearchRepositoryJdbi, since it
was created to match SearchRepositoryWithCountJdbi, and suffers from the
same problems of abstraction.