Add support for String entity IDs
Previously, the EntityId interface had only a single implementor,
UuidEntityId. Although classes in Liflig Document Store said that they
accept any EntityId implementation, it assumed that the IDs would be
UUIDs under the hood. This commit adds a new EntityId implementation,
StringEntityId, and makes uses of EntityId generic so as to no longer
assume that IDs are UUIDs.