Make getHandle public
There are cases where the classes provided by Liflig Document Store are
not sufficient, and one must create a custom SQL statements. In these
cases, one wants to use the same mechanism as Liflig Document Store
for getting a transaction handle, if we are in a transaction. This is
what the getHandle function provides, so it makes sense to make it
public.