Add migrateEntityType
This is a function that we have considered adding previously, allowing a
more powerful kind of migration to change the type of the entity in a
table. However, this migration also has its own pitfalls, and since it's
not clear that it's something we _need_, we're not adding it yet. But
we'll keep it around in a branch in case the need arises in the future.