Add list return to batchCreate/batchUpdate
Now that we support database-generated IDs, it is useful to get this
ID returned from batchCreate. Previously, we did not return results here
out of fear of allocating too much, but I now believe this is a
premature optimization. If we do see cases where this causes memory
issues, we can consider adding specialized versions of the batch methods
that do not return results.