Change transactional to inline function
Since this function takes a lambda parameter, making it inline allows
callers to use non-local returns in the lambda, which is useful for
these types of functions.

This change is technically breaking, since inline functions cannot be
virtual, so we had to remove the `transactional` method from the
`Repository` interface. But we deem this OK, because:

1) A cursory glance at users of the library shows that most use
`RepositoryJdbi` directly, not going through the `Repository`
interface.

2) The `transactional` method on the `Repository` interface was
introduced relatively recently, so it's less likely that users depend
on it.