Remove retry logic from transactional
Transaction serialization failures, which are the type of failures that
can typically be resolved by retrying, can only occur with
REPEATABLE_READ or SERIALIZABLE transaction isolation levels:
https://www.postgresql.org/docs/17/mvcc-serialization-failure-handling.html

The default transaction isolation level in JDBI is READ_COMMITTED, where
serialization failures cannot occur. We do not expose a way to specify
the isolation level to use with `transactional`, so the isolation level
can only be different if the library user has changed the global default
isolation level on their JDBI instance. Since this is not the common
case, we don't want to take on the additional complexity of dealing with
retries in Document Store. Users that have set a custom default
isolation level can handle serialization failure retries themselves.

If we in the future want to improve support for stricter transaction
isolation levels in Document Store, we should revisit this (look at the
previous commits to this one, and consider re-adding the retry logic -
this would require more extensive test coverage).