Move streamAll method from Repository interface to RepositoryJdbi
This method takes a lambda, so we may want to make it `inline` in the
future. This would not work if the method is defined on an interface,
since inline functions cannot be virtual. Since we're about to bump
the major version of the library now anyway (because of changes to
`transactional`), we want to take this opportunity to move this method
off the interface, to make it easier to make it `inline` later if we
want to. Since this method is relatively new anyway, it is unlikely that
any of our users depend on it on the Repository interface specifically.