Rename Versioned.item -> data
When using Liflig Document Store, all tables must have a 'data' column
for the entity JSON data. When fetching entities using a `Repository`,
entities are wrapped in a `Versioned` object with additional metadata
from the table (`version`, `createdAt`, `modifiedAt`).

One would think that the entity data field on `Versioned` would mirror
the column name in the table, 'data' - but no, it's called 'item'. This
creates a discrepancy between how we use entities in the code, versus
how they look in the database (and how we interact with the entity data
in our SQL queries, e.g. `RepositoryJdbi.getByPredicate`).

So to remove this discrepancy, we now rename `Versioned.item` -> `data`.
We keep a property called `item` around for backwards compatibility,
with a `@Deprecated` annotation to ease migration.